--- app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,35 +1,0 @@
-package com.techcourse;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.Request;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ManualHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(ManualHandlerMapping.class);
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    @Override
-    public void initialize() {
-        log.info("Initialized Handler Mapping!");
-        controllers.keySet()
-                .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
-    }
-
-    @Override
-    public boolean hasMapping(final Request request) {
-        return controllers.containsKey(request.getUri());
-    }
-
-    @Override
-    public Controller getHandler(final Request request) {
-        log.debug("Request Mapping Uri : {}", request.getUri());
-        return controllers.get(request.getUri());
-    }
-}
