+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/DispatcherServletInitializer.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet;
 
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
