--- mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adapter/ManualHandlerAdapter.java
@@ -1,25 +1,0 @@
-package webmvc.org.springframework.web.servlet.mvc.adapter;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-public class ManualHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean isHandlerAdapter(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(final Object handler, final HttpServletRequest request, final HttpServletResponse response) {
-        try {
-            final String viewName = ((Controller) handler).execute(request, response);
-            return new ModelAndView(new JspView(viewName));
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
