+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JsonView.java
@@ -5,7 +5,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import web.org.springframework.http.MediaType;
-import webmvc.org.springframework.web.servlet.View;
 
 import java.io.PrintWriter;
 import java.util.Map;
