+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JspView.java
@@ -4,7 +4,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.View;
 
 import java.util.Map;
